requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.0.1
pyotp==2.9.0
qrcode==7.4.2
python-dotenv==1.0.1
bcrypt==4.1.2 